pyproject.toml
[tool.poetry]
name = "undertide"
version = "0.9.7"
description = "data reporting tool"
authors = ["John Cassil <johnpaul@hey.com>"]

[tool.poetry.dependencies]
python = ">=3.11.01,<4"
keyring = "^23.13.1"
"keyrings.google-artifactregistry-auth" = "^1.1.1"
google-cloud-bigquery = "^3.9.0"
pandas = "^1.3.4"
click = "^8.1.3"
loguru = "^0.5.3"
google-cloud-storage = "^2.8.0"
google-cloud-secret-manager = "^2.16.1"
numpy = "^1.22.4"
bs4 = "^0.0.1"
sendgrid = "^6.9.7"
google-cloud-kms = "^2.16.1"
paramiko = "^2.8.0"
tenacity = "^8.2.2"
boto3 = '^1.26.104'
botocore = "^1.29.104"
flask = "^2.2.3"
pyyaml = "^6.0"
pyarrow = "^11.0.0"
snowflake-connector-python = "^3.0.2"
duckdb = "^0.7.1"
jinja2 = "^3.1.2"
mysql-connector-python = "8.0.32"
psycopg2 = "^2.9.5"


[tool.poetry.dev-dependencies]
ipython = "^7.27.0"
pytest = "^7.1.1"
pytest-cov = "^3.0.0"
python-dotenv = "^0.19.2"
black = "^23.3.0"
ruff = "^0.0.260"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra --cov-report=term-missing --cov=src/reports/customer_report --cov=src/util"
testpaths = ["tests"]